Handle fragmented MiiMaker Mii lists
MiiMaker uses 100 slots to store its Miis. Creating a new Mii inserts it
at the first free slot,  while deleting a Mii zero-fills its slot.

It's thus possible for Miis to not be stored seqentially, but with empty
slots in between them, e.g. after deleting Charlie:

 slot | Mii
------+-------------
    0 | Alice
    1 | Bob
    2 | <all zeros>
    3 | Dave

Until now we stopped reading the database at the first empty slot, which
certainly is not the expected behaviour. This commit fixes that by look-
ing for the last occupied slot and then skipping every empty slot that's
in between.